- Resolve TradeExecutionEngine merge conflict on the same axis as
  earlier phases (keep main's BigDecimal TradingParameters, the phase2
  risk wiring, and the phase4 MTF additions).
- HigherTimeframeFilter converts daily-close to double at the boundary.
- MultiTimeframeRouterTest builds Candles with BigDecimal fields.

+++ src/main/java/com/example/cryptobot/strategy/mtf/HigherTimeframeFilter.java
@@ -54,7 +54,7 @@ public final class HigherTimeframeFilter {
         // 1. daily EMA(trendEmaPeriod) trend filter
         if (dailyCandles != null && dailyCandles.size() >= trendEmaPeriod) {
             double ema = Indicators.ema(Indicators.closes(dailyCandles), trendEmaPeriod);
-            double lastClose = dailyCandles.get(dailyCandles.size() - 1).getClosePrice();
+            double lastClose = dailyCandles.get(dailyCandles.size() - 1).getClosePrice().doubleValue();
             if (!Double.isNaN(ema) && lastClose < ema) {
                 return Decision.block("daily close " + lastClose + " < EMA" + trendEmaPeriod + " " + ema);
             }
